% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aictab.R
\name{aictab}
\alias{aictab}
\title{Returns a table with AICs or AICcs, and Akaike weights of a set of models}
\usage{
aictab(...)
}
\arguments{
\item{...}{A set of models of class \code{lm} or \code{merMod}.}
}
\description{
Returns a table with AICs or AICcs (depending on the number of parameters
and the number of observations, see \code{?aic}), and Akaike weights of a set of models.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ cyl, mtcars)
mod2 <- lm(mpg ~ cyl + vs, mtcars)
mod3 <- lm(mpg ~ cyl * vs, mtcars)
aictab(mod1, mod2, mod3)

}
\references{
Burnham, K. P., \& Anderson, D. R. (2002). Model Selection
and Multimodel Inference: A Practical Information-Theoretical Approach.
2d ed. New York: Springer-Verlag.

Burnham, K. P., \& Anderson, D. R. (2004). Multimodel
inference: Understanding AIC and BIC in model selection. Sociological
Methods and Research, 33(2), 261-304.
}
