% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootERseq.R
\name{bootERseq}
\alias{bootERseq}
\title{Sequential evidence ratios of original sample and random rearrangments.}
\usage{
bootERseq(mod1, mod2, samplecol, order_nb, nmin = 10, replace = FALSE)
}
\arguments{
\item{mod1}{A mathematical model, of class "lm" or "lmerMod".}

\item{mod2}{A mathematical model, of class "lm" or "lmerMod" (of the same class of mod1).}

\item{samplecol}{Name of the subject/observation column of your dataframe, as a character vector.
"samplecol" has to be a column of the passed as the "data" argument of the mod1 and mod2 calls.}

\item{order_nb}{Number of random rearrangments to evaluate.}

\item{nmin}{Minimum sample size from which start to compute iterative evidence ratios (ER).}

\item{replace}{If TRUE, corresponds to bootstrap with replacement.}
}
\description{
\code{bootERseq} computes evidence ratios (ER) as a function of sample size,
for a given data set and for N random rearrangments of this dataset.
}
\examples{
library(lme4)
data <- sleepstudy
mod1 <- lm(Reaction ~ 1, data)
mod2 <- lm(Reaction ~ Days, data)
bootERseq(mod1, mod2, samplecol = "Subject", order_nb = 10, nmin = 10, replace = FALSE)

}

