% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compER.R
\name{compER}
\alias{compER}
\title{Simulates sequential testing procedures, using either evidence ratios computed
from WAIC or LOOIC with pseudo-BMA weights or Bayes Factors, for an independent
two-groups comparison}
\usage{
compER(cohensd = 0, nmin = 20, nmax = 100, boundary = 20, B = 12,
  cores = 2)
}
\arguments{
\item{cohensd}{Expected effect size}

\item{nmin}{Minimum sample size from which start computing ERs}

\item{nmax}{Maximum sample size}

\item{boundary}{The Evidence Ratio or the Bayes Factor (resp. its reciprocal)
where the run is stopped as well.}

\item{B}{Number of bootstrap samples (should be dividable by cores)}

\item{cores}{number of parallel processes. If cores is set tp 1, no parallel framework is used.}
}
\description{
Simulates sequential testing procedures, using either evidence ratios computed
from WAIC or LOOIC with pseudo-BMA weights or Bayes Factors, for an independent
two-groups comparison as a function of sample size and standardized mean
difference (Cohen's d). Bayes Factors are computed either via the Savage-Dickey
method (BF_SD) or via bridge sampling (BF_BS).
}
\examples{
\dontrun{
sim <- compER(cohensd = 0, nmin = 20, nmax = 100, boundary = 20, B = 20, cores = 4)
plot(sim)
}

}
\seealso{
\code{\link{simER}}, \code{\link{distER}}
}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
