% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distER.R
\name{distER}
\alias{distER}
\title{Sequential ER for independant two-groups comparisons, as a function of sample size and Cohen's d.}
\usage{
distER(cohensd = 0, n = 100, nmin = 20, nSims = 20)
}
\arguments{
\item{cohensd}{Expected effect size.}

\item{n}{Sample size.}

\item{nmin}{Minimum sample size from which start to compute ER.}

\item{nSims}{Number of experiments to simulate.}
}
\description{
\code{distER} computes evidence ratios (ER) as a function of sample size and cohen's d
}
\examples{
library(ESTER)
distER(cohensd = 0.2, n = 100, nmin = 20, nSims = 20)

}

