% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESTER.R
\docType{package}
\name{ESTER}
\alias{ESTER}
\alias{ESTER-package}
\title{Efficient Sequential Testing with Evidence Ratios}
\description{
The \pkg{ESTER} package implements sequential testing based on evidence
ratios computed from the Akaike weights of a set of models. These weights
are being computed using either the Akaike Information Criterion (AIC) or the
Bayesian Information Criterion (BIC).
}
\details{
See \code{vignette("ESTER")} for a general introduction and overview.
}
\seealso{
\code{\link{ictab}}, \code{\link{simER}}, \code{\link{seqER}}
}
\author{
Ladislas Nalborczyk \cr
Maintainer: Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
