% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqERboot.R
\name{seqERboot}
\alias{seqERboot}
\title{Sequential evidence ratios of original sample and random rearrangments.}
\usage{
seqERboot(mod1, mod2, samplecol = NULL, order_nb, nmin = 10,
  replace = FALSE)
}
\arguments{
\item{mod1}{A mathematical model, of class "lm" or "lmerMod".}

\item{mod2}{A mathematical model, of class "lm" or "lmerMod" (of the same class of mod1).}

\item{samplecol}{If applicable, name of the subject column of your dataframe, as a character vector.
"samplecol" has to be a column of the data passed as the "data" argument of the mod1 and mod2 calls (default is NULL).}

\item{order_nb}{Number of random rearrangments to evaluate.}

\item{nmin}{Minimum sample size from which start to compute iterative evidence ratios (ER).}

\item{replace}{If TRUE, corresponds to bootstrap with replacement.}
}
\description{
\code{seqERboot} computes evidence ratios (ER) as a function of sample size,
for a given data set and for N random rearrangments of this dataset.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ 1, mtcars)
mod2 <- lm(mpg ~ cyl, mtcars)
seq_boot_mtcars <- seqERboot(mod1, mod2, order_nb = 10, nmin = 10, replace = FALSE)

}

