% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simER.R
\name{simER}
\alias{simER}
\title{Sequential ER for independant two-groups comparisons, as a function of sample size and Cohen's d.}
\usage{
simER(cohensd = 0, n = 100, nmin = 20, nmax = 200)
}
\arguments{
\item{cohensd}{Expected effect size.}

\item{n}{Sample size.}

\item{nmin}{Minimum sample size from which start to compute ER.}

\item{nmax}{Maximum sample size to which stop to compute ER.}
}
\description{
\code{simER} computes evidence ratios (ER) as a function of sample size and cohen's d
}
\examples{
library(ESTER)
simER(cohensd = 0.2, n = 100, nmin = 20, nmax = 1000)

}

