% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticer.R
\docType{package}
\name{sticer}
\alias{sticer}
\alias{sticer-package}
\title{Sequential Testing with Information Criteria and Evidence Ratios}
\description{
The \pkg{sticer} package implements sequential testing based on evidence
ratios computed from the Akaike weights of a set of models. These weights
are being computed using either the Akaike Information Criterion (AIC) or the
Bayesian Information Criterion (BIC).
}
\details{
See \code{vignette("sticer")} for a general introduction and overview.
}
\seealso{
\code{\link{ictab}}, \code{\link{simER}}, \code{\link{seqER}}
}
\author{
Ladislas Nalborczyk \cr
Maintainer: Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
