% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Computing likelihood and posterior probabilities of population effect sizes
from simulations ran with \code{simER}}
\usage{
bayes(sim, type = "posterior", threshold, prior = NULL, palette = "PuBu")
}
\arguments{
\item{sim}{A \code{simER} object.}

\item{type}{Should we plot the likelihood or the posterior, or should we output a table ?}

\item{threshold}{A logical statement (see examples)}

\item{prior}{Defining the prior on effect sizes (by default uniform prior)}

\item{palette}{Defining the color palette to be used for plots.}
}
\value{
Either a plot of the evolution of the likelihood or the posterior
probability, or a dataframe summarising these probabilities for each effect size
and sample size.
}
\description{
Computing the likelihood of observing an evidence ratio above or below
a certain threshold, given a certain effect size and a sample size,
from simulations ran with \code{simER}. Then, using Bayes theorem, we
can compute the inverse probability (the posterior probability) of
the population effect size given a certain evidence ratio and
sample size (and prior probability).
}
\examples{
\dontrun{
# loading some simulated data
load(url("https://github.com/lnalborczyk/shiny/blob/master/sims.RData?raw=true") )

# plotting the likelihood
bayes(sim, type = "likelihood", threshold = ER > 10)

# plotting the posterior probability
bayes(sim, type = "posterior", threshold = ER > 10)

# changing priors
pr <- c(0.1, 0.1, 0.1, 0.2, 0.2, 0.15, 0.05, 0.05, 0.025, 0.025)
bayes(sim, type = "posterior", threshold = ER > 10, prior = pr)

# obtaining the probability of observing an ER > 10 if the
# population effect size is equal to 0 and n = 113 (with uniform priors)
bayes(sim, type = "table", threshold = ER > 10) \%>\%
filter(n == 113 & true.ES == 0)

# plotting the prior and posterior distributions on effect sizes

bayes(sim, type = "table", threshold = ER > 10) \%>\%
filter(n == 113) \%>\% # as an example
gather(type, prob, likelihood:posterior) \%>\%
filter(type \%in\% c("prior", "posterior") ) \%>\%
ggplot(aes(x = true.ES, y = prob, fill = type) ) +
geom_bar(stat = "identity", position = position_dodge() ) +
theme_bw(base_size = 12) +
scale_fill_grey(start = 0.25, end = 0.75) +
xlab(expression(delta) ) +
ylab("")
}

}
\seealso{
\code{\link{simER}}
}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
