% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compER.R
\name{compER}
\alias{compER}
\title{Simulates sequential testing procedures, using either evidence ratios computed
from WAIC or LOOIC with pseudo-BMA weights or Bayes Factors, for an independent
two-groups comparison}
\usage{
compER(cohensd = 0, nmin = 20, nmax = 100, boundary = 10,
  prior = "normal(0, 5)", B = 20, cores = 2)
}
\arguments{
\item{cohensd}{Expected effect size}

\item{nmin}{Minimum sample size from which start computing ERs}

\item{nmax}{Maximum sample size}

\item{boundary}{The Evidence Ratio or the Bayes Factor (or its reciprocal)
to which the run is stopped as well}

\item{prior}{The prior that sould be placed on the slope. This will strongly
influence the Bayes Factor computed using the Savage-Dickey method (BF_SD).}

\item{B}{Number of bootstrap samples (should be dividable by cores)}

\item{cores}{Number of parallel processes. If cores is set to 1, no parallel framework is used.}
}
\value{
An object of class \code{data.frame}, which contains the value of the
evidence ratio (either WAIC_ER or LOO_ER) or the Bayes Factor (either DF_SD or BF_BS),
at a specific sample size n, for each simulation id.
}
\description{
Simulates sequential testing procedures, using either evidence ratios computed
from WAIC or LOOIC with pseudo-BMA weights or Bayes Factors, for an independent
two-groups comparison as a function of sample size and standardized mean
difference (Cohen's d). Bayes Factors are computed either via the Savage-Dickey
method (BF_SD) or via bridge sampling (BF_BS).
}
\examples{
\dontrun{
sim <- compER(cohensd = 0.8, nmin = 20, nmax = 100, B = 20, cores = 4)
plot(sim)
}

}
\seealso{
\code{\link{simER}}
}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
