% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distER.R
\name{distER}
\alias{distER}
\title{Simulates many sequential evidence ratios, keeps the last of each simulation,
and plot their distribution.}
\usage{
distER(cohensd, nmin, n, nsims)
}
\arguments{
\item{cohensd}{Expected effect size.}

\item{nmin}{Minimum sample size from which start to compute ER.}

\item{n}{Sample size.}

\item{nsims}{Number of experiments to simulate.}
}
\description{
Simulates many sequential evidence ratios, keeps the last of each simulation,
and plot their distribution.
}
\examples{
library(ESTER)
ER <- distER(cohensd = 0.6, n = 100, nmin = 20, nsims = 100)

}
