% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itERrand.R
\name{itERrand}
\alias{itERrand}
\title{Iterative evience ratios (ER) of original sample and random rearrangments}
\usage{
itERrand(data, mod1, mod2, samplecol, order_nb, nmin = 10)
}
\arguments{
\item{mod1}{should be a mathematical model, of class "lm" or "lmerMod"}

\item{mod2}{should be a mathematical model, of class "lm" or "lmerMod" (of the same class of mod1)}

\item{samplecol}{should be the name of the participant/subject column of your dataframe, as a character vector}

\item{nmin}{is the minimum sample size to which start to compute iterative evidence ratios (ER)}
}
\description{
Iterative evience ratios (ER) of original sample and random rearrangments
}
\examples{
library(lme4)
data <- sleepstudy
mod1 <- lm(Reaction ~ 1, data)
mod2 <- lm(Reaction ~ Days, data)
iER(dat, mod1, mod2, "Subject", 10)

}

