% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simER.R
\name{plot.simER}
\alias{plot.simER}
\title{Plotting the results of \code{simER}}
\usage{
\method{plot}{simER}(x, log = TRUE, hist = FALSE, ...)
}
\arguments{
\item{x}{Should be a \code{simER} object}

\item{log}{Should the y-axis be log-transformed ?}

\item{hist}{Should plot the histogram of simulations hitting either the lower,
the upper boundary or stopping at nmax ?}

\item{...}{Additional parameters}
}
\description{
Plotting the results of \code{simER}.
}
\author{
Ladislas Nalborczyk <\email{ladislas.nalborczyk@gmail.com}>
}
