% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqER.R
\name{seqER}
\alias{seqER}
\title{Computes sequential evidence ratios}
\usage{
seqER(ic, mod1, mod2, nmin, id = NULL)
}
\arguments{
\item{ic}{Indicates whether to use the aic or the bic.}

\item{mod1}{A model of class \code{lm} or \code{lmerMod}.}

\item{mod2}{A model of class \code{lm} or \code{lmerMod} (of the same class of mod1).}

\item{nmin}{Minimum sample size from which start to compute sequential evidence ratios.}

\item{id}{If applicable (e.g., repeated measures), name of the id column of your
dataframe, in character string.}
}
\description{
Computes sequential evidence ratios.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ cyl, mtcars)
mod2 <- lm(mpg ~ cyl + disp, mtcars)
seq_mtcars <- seqER(ic = aic, mod1, mod2, nmin = 10)

# Example with repeated measures
library(lme4)
data(sleepstudy)
mod1 <- lm(Reaction ~ Days, sleepstudy)
mod2 <- lm(Reaction ~ Days + I(Days^2), sleepstudy)
seqER(ic = aic, mod1, mod2, nmin = 10, id = "Subject")

}
