% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqER.R
\name{seqER}
\alias{seqER}
\title{Computes sequential evidence ratios}
\usage{
seqER(mod1, mod2, nmin, samplecol = NULL)
}
\arguments{
\item{mod1}{Amodel of class "lm" or "lmerMod".}

\item{mod2}{A model of class "lm" or "lmerMod" (of the same class of mod1).}

\item{nmin}{Minimum sample size from which start to compute sequential evidence ratios.}

\item{samplecol}{If applicable (e.g., repeated measures), name of the subject column of your
dataframe.}
}
\description{
Computes sequential evidence ratios.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ cyl, mtcars)
mod2 <- lm(mpg ~ cyl + disp, mtcars)
seq_mtcars <- seqER(mod1, mod2, nmin = 10)

}
