% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqER.R
\name{seqER}
\alias{seqER}
\title{Sequential evidence ratios.}
\usage{
seqER(mod1, mod2, samplecol, nmin)
}
\arguments{
\item{mod1}{A mathematical model, of class "lm" or "lmerMod".}

\item{mod2}{A mathematical model, of class "lm" or "lmerMod" (of the same class of mod1).}

\item{samplecol}{Name of the subject/observation column of your dataframe, as a character vector.
"samplecol" has to be a column of the passed as the "data" argument of the mod1 and mod2 calls.}

\item{nmin}{Minimum sample size from which start to compute iterative evidence ratios (ER).}
}
\description{
\code{seqER} computes evidence ratios (ER) as a function of sample size, for a given data set.
}
\examples{
library(lme4)
data <- sleepstudy
mod1 <- lm(Reaction ~ 1, data)
mod2 <- lm(Reaction ~ Days, data)
seqER(mod1, mod2, samplecol = "Subject", nmin = 10)

}

