% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqER.R
\name{seqER}
\alias{seqER}
\title{Sequential evidence ratios.}
\usage{
seqER(mod1, mod2, samplecol = NULL, nmin)
}
\arguments{
\item{mod1}{A mathematical model, of class "lm" or "lmerMod".}

\item{mod2}{A mathematical model, of class "lm" or "lmerMod" (of the same class of mod1).}

\item{samplecol}{If applicable (e.g., repeated measures), name of the subject column of your dataframe, as a character vector.}

\item{nmin}{Minimum sample size from which start to compute iterative evidence ratios (ER).}
}
\description{
\code{seqER} computes evidence ratios (ER) as a function of sample size, for a given data set.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ cyl, mtcars)
mod2 <- lm(mpg ~ cyl + disp, mtcars)
seq_mtcars <- seqER(mod1, mod2, nmin = 10)

}

