% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqERboot.R
\name{seqERboot}
\alias{seqERboot}
\title{Computes sequential evidence ratios for a given data set and bootstrapped samples}
\usage{
seqERboot(mod1, mod2, nmin, id = NULL, order_nb, replace = FALSE)
}
\arguments{
\item{mod1}{A model of class "lm" or "lmerMod".}

\item{mod2}{A model of class "lm" or "lmerMod" (of the same class of mod1).}

\item{nmin}{Minimum sample size from which start to compute sequential evidence ratios.}

\item{id}{If applicable (e.g., repeated measures), name of the id column of your
dataframe, in character string.}

\item{order_nb}{Number of random rearrangments to evaluate.}

\item{replace}{If TRUE, corresponds to bootstrap with replacement.}
}
\description{
Computes sequential evidence ratios for a given data set as well as
for order_nb random rearrangments of this dataset.
}
\examples{
data(mtcars)
mod1 <- lm(mpg ~ cyl, mtcars)
mod2 <- lm(mpg ~ cyl + disp, mtcars)
seq_boot_mtcars <- seqERboot(mod1, mod2, nmin = 10, order_nb = 20, replace = FALSE)

# Example with repeated measures
library(lme4)
data(sleepstudy)
mod1 <- lm(Reaction ~ Days, sleepstudy)
mod2 <- lm(Reaction ~ Days + I(Days^2), sleepstudy)
seqERboot(mod1, mod2, nmin = 10, id = "Subject", order_nb = 20, replace = TRUE)

}
