% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simER.R
\name{simER}
\alias{simER}
\title{Simulates a sequential ER for independant two-groups comparisons}
\usage{
simER(cohensd, nmin, n, plot = TRUE)
}
\arguments{
\item{cohensd}{Expected effect size}

\item{nmin}{Minimum sample size from which start computing ERs}

\item{n}{Total sample size}

\item{plot}{If TRUE, produces a plot of the evolution of the ERs}
}
\description{
Simulates a sequential ER for independant two-groups comparisons,
as a function of sample size and effect size (Cohen's d).
}
\examples{
simER(cohensd = 0.6, nmin = 20, n = 100, plot = TRUE)

}
